requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.2
typing_extensions==4.12.2
